refactor the app bycompingin the routes under one file.
