+++ app.py
@@ -1,15 +1,8 @@
-from flask import Flask, render_template
+from flask import Flask
 
 app = Flask(__name__)
 
-@app.route('/')
-@app.route('/index')
-def index():
-    return render_template('index.html', title='Home')
-
-@app.route('/about')
-def about():
-    return render_template('about.html', title='about')
+from routes import *
 
 if __name__ == "__main__":
     app.run(debug=True)
